Modified the file structure of components
