+++ doc/confapp_flutter/lib/components/components.dart
@@ -1,0 +1,2 @@
+export './rotate_generator.dart';
+export './sample_generator.dart';
