+++ doc/confapp_flutter/lib/hcl_components.dart
@@ -1,0 +1,9 @@
+import 'package:confapp_flutter/components/config.dart';
+import 'package:confapp_flutter/components/components.dart';
+
+class WebPageGenerator {
+  final List<ConfigGenerator> generators = [
+    RotateGenerator(),
+    SampleGenerator()
+  ];
+}
